Split backend and frontend containers
