+++ backend/package.json
@@ -7,7 +7,8 @@
   "scripts": {
     "build": "tsc",
     "start": "node out/index.js",
-    "watch": "tsc-watch --onSuccess \"node out/index.js\""
+    "watch": "tsc-watch --onSuccess \"node out/index.js\"",
+    "profile": "node --prof out/index.js"
   },
   "keywords": [
     "sync",
