+++ nginx.conf
@@ -1,0 +1,26 @@
+server {
+    listen 8001 ;
+    server_name localhost ;
+    root /usr/share/nginx/html ;
+    index index.html ;
+
+    location / {
+        try_files $uri $uri/ /index.html ;
+    }
+
+    location /socket.io/ {
+        proxy_pass http://backend:8080 ;
+        proxy_set_header Upgrade $http_upgrade ;
+        proxy_set_header Connection "upgrade";
+    }
+
+    location /youtube/ {
+        proxy_pass http://backend:8080 ;
+        proxy_set_header Host $host ;
+    }
+
+    location /cdn/ {
+        proxy_pass http://backend:8080 ;
+        proxy_set_header Host $host ;
+    }
+}
